+++ src/kra_etims/__init__.py
@@ -140,7 +140,7 @@ from .exceptions import (
     ZReportAlreadyIssuedError,
 )
 
-__version__ = "0.6.0"
+__version__ = "0.6.1"
 
 __all__ = [
     # Clients